+++ src/features/post/public/postList/store/PostSearchStateStore.tsx
